- finalize optimization
