--- BallisticCalculator.Benchmark/resources/g1_nowind.txt
@@ -1,27 +1,0 @@
-﻿ammo;0.365G1;65gr;2600ft/s
-rifle;2.50in;100yd
-wind;0mph;0°
-atmosphere;59.0°F;78;29.92inHg;0ft
-shot;0°;0°
-yd;in;moa;in;moa;ft/s;mach;ft-lb;s;in;moa
-0;-2.5;0.0;0.0;0.0;2600.0;2.329;975.5;0.000;0.0;0.0
-50;-0.5;-1.0;0.0;0.0;2478.7;2.220;886.6;0.059;0.0;0.0
-100;-0.0;-0.0;0.0;0.0;2360.6;2.114;804.1;0.121;0.0;0.0
-150;-1.0;-0.6;0.0;0.0;2245.7;2.011;727.7;0.186;0.0;0.0
-200;-3.8;-1.8;0.0;0.0;2133.9;1.911;657.1;0.255;0.0;0.0
-250;-8.4;-3.2;0.0;0.0;2025.3;1.814;591.9;0.327;0.0;0.0
-300;-15.2;-4.8;0.0;0.0;1920.1;1.720;532.0;0.403;0.0;0.0
-350;-24.3;-6.6;0.0;0.0;1818.4;1.629;477.2;0.483;0.0;0.0
-400;-36.1;-8.6;0.0;0.0;1720.5;1.541;427.2;0.568;0.0;0.0
-450;-50.8;-10.8;0.0;0.0;1626.8;1.457;381.9;0.658;0.0;0.0
-500;-68.8;-13.1;0.0;0.0;1537.5;1.377;341.1;0.753;0.0;0.0
-550;-90.4;-15.7;0.0;0.0;1453.3;1.302;304.8;0.853;0.0;0.0
-600;-116.2;-18.5;0.0;0.0;1374.5;1.231;272.6;0.959;0.0;0.0
-650;-146.6;-21.5;0.0;0.0;1301.8;1.166;244.6;1.071;0.0;0.0
-700;-182.1;-24.8;0.0;0.0;1235.8;1.107;220.4;1.190;0.0;0.0
-750;-223.3;-28.4;0.0;0.0;1177.2;1.054;200.0;1.314;0.0;0.0
-800;-270.8;-32.3;0.0;0.0;1126.2;1.009;183.0;1.444;0.0;0.0
-850;-325.1;-36.5;0.0;0.0;1082.6;0.970;169.1;1.580;0.0;0.0
-900;-386.9;-41.1;0.0;0.0;1045.3;0.936;157.7;1.722;0.0;0.0
-950;-456.6;-45.9;0.0;0.0;1013.1;0.907;148.1;1.868;0.0;0.0
-1000;-534.8;-51.1;0.0;0.0;984.7;0.882;139.9;2.018;0.0;0.0
