--- BallisticCalculator.Benchmark/resources/g1_nowind_up.txt
@@ -1,28 +1,0 @@
-﻿ammo;0.365G1;65gr;2600ft/s
-rifle;2.50in;100yd
-wind;0mph;0°
-atmosphere;59.0°F;78;29.92inHg;0ft
-shot;10°;0°
-yd;in;moa;in;moa;ft/s;mach;ft-lb;s;in;moa
-0;-2.5;0.0;0.0;0.0;2600.0;2.329;975.5;0.000;0.0;0.0
-50;-0.5;-1.0;0.0;0.0;2478.4;2.220;886.4;0.059;0.0;0.0
-100;0.0;0.0;0.0;0.0;2360.1;2.114;803.8;0.121;0.0;0.0
-150;-0.9;-0.6;0.0;0.0;2245.1;2.011;727.3;0.186;0.0;0.0
-200;-3.6;-1.7;0.0;0.0;2133.2;1.911;656.7;0.255;0.0;0.0
-250;-8.1;-3.1;0.0;0.0;2024.7;1.814;591.5;0.327;0.0;0.0
-300;-14.8;-4.7;0.0;0.0;1919.5;1.720;531.7;0.403;0.0;0.0
-350;-23.7;-6.5;0.0;0.0;1817.9;1.629;476.9;0.483;0.0;0.0
-400;-35.3;-8.4;0.0;0.0;1720.1;1.542;427.0;0.568;0.0;0.0
-450;-49.7;-10.6;0.0;0.0;1626.5;1.458;381.8;0.658;0.0;0.0
-500;-67.4;-12.9;0.0;0.0;1537.4;1.378;341.1;0.753;0.0;0.0
-550;-88.8;-15.4;0.0;0.0;1453.2;1.303;304.7;0.853;0.0;0.0
-600;-114.2;-18.2;0.0;0.0;1374.5;1.232;272.6;0.959;0.0;0.0
-650;-144.1;-21.2;0.0;0.0;1301.8;1.167;244.6;1.072;0.0;0.0
-700;-179.1;-24.4;0.0;0.0;1235.8;1.108;220.4;1.190;0.0;0.0
-750;-219.7;-28.0;0.0;0.0;1176.9;1.056;199.9;1.314;0.0;0.0
-800;-266.6;-31.8;0.0;0.0;1125.7;1.010;182.9;1.445;0.0;0.0
-850;-320.2;-36.0;0.0;0.0;1081.8;0.970;168.9;1.581;0.0;0.0
-900;-381.2;-40.4;0.0;0.0;1044.2;0.937;157.3;1.722;0.0;0.0
-950;-450.1;-45.2;0.0;0.0;1011.6;0.908;147.7;1.868;0.0;0.0
-1000;-527.4;-50.4;0.0;0.0;982.8;0.882;139.4;2.019;0.0;0.0
-
