--- BallisticCalculator.Benchmark/resources/g1_twist.txt
@@ -1,27 +1,0 @@
-﻿ammo;0.5G1;220gr;3006.6ft/s;0.308in;1.5in
-rifle;1.50in;100yd;12in;right
-wind;0mph;0°
-atmosphere;59.0°F;78;29.92inHg;0ft
-shot;0°;0°
-yd;in;moa;in;moa;ft/s;mach;ft-lb;s;in;moa
-0;-1.5;0.0;0.0;0.0;3006.6;2.693;4415.0;0.000;0.0;0.0
-50;-0.2;-0.5;-0.01;0.0;2909.1;2.606;4133.3;0.051;8.9;17.0
-100;-0.0;-0.0;-0.05;0.0;2813.8;2.520;3866.9;0.103;18.2;17.3
-150;-0.9;-0.5;-0.1;-0.1;2720.5;2.437;3614.8;0.157;27.7;17.6
-200;-2.9;-1.4;-0.2;-0.1;2629.2;2.355;3376.2;0.213;37.6;17.9
-250;-6.2;-2.4;-0.3;-0.1;2539.7;2.275;3150.3;0.272;47.8;18.3
-300;-10.8;-3.4;-0.4;-0.1;2452.0;2.196;2936.4;0.332;58.4;18.6
-350;-16.9;-4.6;-0.6;-0.2;2366.0;2.119;2734.1;0.394;69.3;18.9
-400;-24.6;-5.9;-0.7;-0.2;2281.7;2.044;2542.7;0.458;80.7;19.3
-450;-33.9;-7.2;-0.9;-0.2;2199.1;1.970;2361.9;0.525;92.5;19.6
-500;-45.0;-8.6;-1.2;-0.2;2118.1;1.897;2191.2;0.595;104.7;20.0
-550;-58.0;-10.1;-1.5;-0.3;2038.9;1.826;2030.4;0.667;117.4;20.4
-600;-73.1;-11.6;-1.8;-0.3;1961.4;1.757;1879.0;0.742;130.6;20.8
-650;-90.5;-13.3;-2.1;-0.3;1885.8;1.689;1737.0;0.820;144.3;21.2
-700;-110.4;-15.1;-2.5;-0.3;1812.1;1.623;1603.9;0.901;158.6;21.6
-750;-132.9;-16.9;-3.0;-0.4;1740.5;1.559;1479.6;0.986;173.5;22.1
-800;-158.2;-18.9;-3.5;-0.4;1671.0;1.497;1363.8;1.074;189.0;22.6
-850;-186.7;-21.0;-4.1;-0.5;1603.8;1.437;1256.3;1.165;205.1;23.0
-900;-218.5;-23.2;-4.7;-0.5;1539.1;1.379;1157.0;1.261;221.9;23.5
-950;-254.1;-25.5;-5.4;-0.5;1477.0;1.323;1065.6;1.360;239.4;24.1
-1000;-293.6;-28.0;-6.2;-0.6;1417.8;1.270;981.8;1.464;257.7;24.6
