--- BallisticCalculator.Benchmark/resources/g1_wind.txt
@@ -1,27 +1,0 @@
-﻿ammo;0.365G1;65gr;2600ft/s
-rifle;2.50in;100yd
-wind;20mph;30°
-atmosphere;59.0°F;78;29.92inHg;0ft
-shot;0°;0°
-yd;in;moa;in;moa;ft/s;mach;ft-lb;s;in;moa
-0;-2.5;0.0;0.0;0.0;2600.0;2.329;975.5;0.000;0.0;0.0
-50;-0.5;-1.0;0.2;0.5;2480.5;2.222;887.9;0.059;0.0;0.0
-100;-0.0;-0.0;1.0;1.0;2364.3;2.118;806.6;0.121;0.0;0.0
-150;-1.0;-0.6;2.3;1.5;2251.2;2.016;731.3;0.186;0.0;0.0
-200;-3.7;-1.8;4.2;2.0;2141.2;1.918;661.6;0.254;0.0;0.0
-250;-8.4;-3.2;6.7;2.6;2034.5;1.822;597.3;0.326;0.0;0.0
-300;-15.1;-4.8;9.9;3.2;1931.1;1.730;538.2;0.402;0.0;0.0
-350;-24.1;-6.6;13.8;3.8;1831.3;1.640;484.0;0.482;0.0;0.0
-400;-35.8;-8.5;18.5;4.4;1735.3;1.554;434.5;0.566;0.0;0.0
-450;-50.3;-10.7;24.1;5.1;1643.4;1.472;389.7;0.655;0.0;0.0
-500;-68.0;-13.0;30.5;5.8;1556.0;1.394;349.4;0.748;0.0;0.0
-550;-89.3;-15.5;37.8;6.6;1473.5;1.320;313.3;0.848;0.0;0.0
-600;-114.6;-18.2;46.2;7.4;1396.5;1.251;281.4;0.952;0.0;0.0
-650;-144.3;-21.2;55.5;8.2;1325.4;1.187;253.5;1.062;0.0;0.0
-700;-178.9;-24.4;65.9;9.0;1260.8;1.129;229.4;1.179;0.0;0.0
-750;-218.9;-27.9;77.3;9.8;1203.3;1.078;208.9;1.300;0.0;0.0
-800;-264.9;-31.6;89.7;10.7;1153.3;1.033;191.9;1.428;0.0;0.0
-850;-317.5;-35.7;103.0;11.6;1110.3;0.994;177.9;1.561;0.0;0.0
-900;-377.0;-40.0;117.2;12.4;1073.5;0.962;166.3;1.698;0.0;0.0
-950;-444.0;-44.6;132.2;13.3;1041.7;0.933;156.6;1.840;0.0;0.0
-1000;-518.9;-49.6;147.9;14.1;1013.6;0.908;148.3;1.986;0.0;0.0
