--- BallisticCalculator.Benchmark/resources/g1_wind_cold.txt
@@ -1,27 +1,0 @@
-﻿ammo;0.465G1;65gr;3200ft/s
-rifle;3.20in;300yd
-wind;15mph;45°
-atmosphere;10.0°F;5;32inHg;0ft
-shot;0°;0°
-yd;in;moa;in;moa;ft/s;mach;ft-lb;s;in;moa
-0;-3.2;0.0;0.0;0.0;3200.0;3.012;1477.7;0.000;0.0;0.0
-50;-0.1;-0.2;0.2;0.3;3072.9;2.892;1362.6;0.048;0.0;0.0
-100;2.1;2.0;0.7;0.7;2949.7;2.776;1255.6;0.098;0.0;0.0
-150;3.3;2.1;1.7;1.1;2830.2;2.664;1155.9;0.150;0.0;0.0
-200;3.4;1.6;3.0;1.5;2714.2;2.555;1063.1;0.204;0.0;0.0
-250;2.3;0.9;4.8;1.8;2601.4;2.449;976.5;0.260;0.0;0.0
-300;-0.0;-0.0;7.1;2.3;2491.6;2.345;895.8;0.319;0.0;0.0
-350;-3.8;-1.0;9.8;2.7;2384.6;2.244;820.5;0.381;0.0;0.0
-400;-9.1;-2.2;13.1;3.1;2280.2;2.146;750.3;0.445;0.0;0.0
-450;-16.0;-3.4;17.0;3.6;2178.6;2.051;684.9;0.512;0.0;0.0
-500;-24.8;-4.7;21.4;4.1;2079.5;1.957;624.0;0.583;0.0;0.0
-550;-35.6;-6.2;26.4;4.6;1983.0;1.866;567.4;0.657;0.0;0.0
-600;-48.6;-7.7;32.2;5.1;1889.2;1.778;515.0;0.734;0.0;0.0
-650;-64.0;-9.4;38.7;5.7;1798.2;1.693;466.6;0.815;0.0;0.0
-700;-82.1;-11.2;45.9;6.3;1710.3;1.610;422.1;0.901;0.0;0.0
-750;-103.2;-13.1;54.0;6.9;1625.6;1.530;381.3;0.991;0.0;0.0
-800;-127.5;-15.2;63.0;7.5;1544.3;1.454;344.1;1.086;0.0;0.0
-850;-155.5;-17.5;72.9;8.2;1466.8;1.381;310.5;1.185;0.0;0.0
-900;-187.5;-19.9;83.8;8.9;1393.5;1.312;280.2;1.290;0.0;0.0
-950;-224.0;-22.5;95.7;9.6;1324.6;1.247;253.2;1.401;0.0;0.0
-1000;-265.4;-25.3;108.7;10.4;1260.7;1.187;229.3;1.517;0.0;0.0
