--- BallisticCalculator.Benchmark/resources/g1_wind_hot.txt
@@ -1,27 +1,0 @@
-﻿ammo;0.465G1;65gr;3200ft/s
-rifle;3.20in;300yd
-wind;15mph;45°
-atmosphere;80.0°F;50;29.5inHg;0ft
-shot;0°;0°
-yd;in;moa;in;moa;ft/s;mach;ft-lb;s;in;moa
-0;-3.2;0.0;0.0;0.0;3200.0;2.810;1477.7;0.000;0.0;0.0
-50;-0.2;-0.3;0.1;0.3;3096.6;2.719;1383.7;0.048;0.0;0.0
-100;1.9;1.8;0.6;0.6;2995.7;2.630;1295.0;0.097;0.0;0.0
-150;3.1;2.0;1.3;0.9;2897.0;2.544;1211.1;0.148;0.0;0.0
-200;3.2;1.5;2.4;1.2;2800.5;2.459;1131.8;0.200;0.0;0.0
-250;2.2;0.8;3.9;1.5;2706.1;2.376;1056.7;0.255;0.0;0.0
-300;-0.0;-0.0;5.7;1.8;2613.6;2.295;985.8;0.311;0.0;0.0
-350;-3.5;-0.9;7.8;2.1;2523.0;2.215;918.6;0.370;0.0;0.0
-400;-8.3;-2.0;10.4;2.5;2434.3;2.137;855.1;0.430;0.0;0.0
-450;-14.6;-3.1;13.4;2.8;2347.3;2.061;795.1;0.493;0.0;0.0
-500;-22.4;-4.3;16.8;3.2;2262.0;1.986;738.4;0.558;0.0;0.0
-550;-32.0;-5.6;20.7;3.6;2178.6;1.913;684.9;0.626;0.0;0.0
-600;-43.4;-6.9;25.0;4.0;2096.9;1.841;634.5;0.696;0.0;0.0
-650;-56.8;-8.3;29.9;4.4;2017.1;1.771;587.1;0.769;0.0;0.0
-700;-72.3;-9.9;35.4;4.8;1939.3;1.703;542.7;0.845;0.0;0.0
-750;-90.1;-11.5;41.4;5.3;1863.4;1.636;501.1;0.924;0.0;0.0
-800;-110.4;-13.2;48.0;5.7;1789.8;1.572;462.2;1.006;0.0;0.0
-850;-133.4;-15.0;55.2;6.2;1718.3;1.509;426.1;1.091;0.0;0.0
-900;-159.3;-16.9;63.2;6.7;1649.3;1.448;392.6;1.180;0.0;0.0
-950;-188.4;-18.9;71.8;7.2;1582.9;1.390;361.6;1.273;0.0;0.0
-1000;-221.0;-21.1;81.1;7.7;1519.3;1.334;333.1;1.370;0.0;0.0
