--- BallisticCalculator.Benchmark/resources/g7_nowind.txt
@@ -1,27 +1,0 @@
-﻿ammo;0.465G7;65gr;3200ft/s
-rifle;3.20in;300yd
-wind;0mph;0°
-atmosphere;59.0°F;78;29.92inHg;0ft
-shot;0°;0°
-yd;in;moa;in;moa;ft/s;mach;ft-lb;s;in;moa
-0;-3.2;0.0;0.0;0.0;3200.0;2.866;1477.7;0.000;0.0;0.0
-50;-0.4;-0.7;0.0;0.0;3146.7;2.818;1428.8;0.047;0.0;0.0
-100;1.6;1.5;0.0;0.0;3093.7;2.771;1381.1;0.095;0.0;0.0
-150;2.6;1.7;0.0;0.0;3041.1;2.724;1334.6;0.144;0.0;0.0
-200;2.7;1.3;0.0;0.0;2988.9;2.677;1289.1;0.194;0.0;0.0
-250;1.9;0.7;0.0;0.0;2937.1;2.631;1244.8;0.245;0.0;0.0
-300;-0.0;-0.0;0.0;0.0;2885.6;2.585;1201.6;0.296;0.0;0.0
-350;-2.9;-0.8;0.0;0.0;2834.6;2.539;1159.5;0.349;0.0;0.0
-400;-6.9;-1.7;0.0;0.0;2784.1;2.494;1118.5;0.402;0.0;0.0
-450;-12.0;-2.6;0.0;0.0;2734.0;2.449;1078.6;0.456;0.0;0.0
-500;-18.3;-3.5;0.0;0.0;2684.4;2.404;1039.8;0.512;0.0;0.0
-550;-25.8;-4.5;0.0;0.0;2635.2;2.360;1002.1;0.568;0.0;0.0
-600;-34.5;-5.5;0.0;0.0;2586.5;2.317;965.4;0.626;0.0;0.0
-650;-44.6;-6.5;0.0;0.0;2538.3;2.274;929.8;0.684;0.0;0.0
-700;-56.0;-7.6;0.0;0.0;2490.7;2.231;895.2;0.744;0.0;0.0
-750;-68.7;-8.8;0.0;0.0;2443.5;2.189;861.6;0.805;0.0;0.0
-800;-83.0;-9.9;0.0;0.0;2396.8;2.147;829.0;0.867;0.0;0.0
-850;-98.7;-11.1;0.0;0.0;2350.6;2.105;797.3;0.930;0.0;0.0
-900;-116.1;-12.3;0.0;0.0;2305.0;2.065;766.7;0.994;0.0;0.0
-950;-135.0;-13.6;0.0;0.0;2259.9;2.024;737.0;1.060;0.0;0.0
-1000;-155.7;-14.9;0.0;0.0;2215.3;1.984;708.2;1.127;0.0;0.0
